While Loop and Do-While Loop
Do-While Loop will allow us to print out the result, then check the
while conditions to see if it is able to pass or not. If it does not
pass, the system of Do-While Loop will STOP looping.

Also, the stop an infinity while loop, add if statement with specific
conditions and break;